Used bigger number (#663)
* Used bigger number

* Use even bigger numbers

+++ grakn-graql/src/main/java/ai/grakn/graql/internal/gremlin/fragment/InSubFragment.java
@@ -38,7 +38,7 @@ class InSubFragment extends AbstractFragment {
     }
 
     @Override
-    public long fragmentCost(long previousCost) {
+    public double fragmentCost(double previousCost) {
         return previousCost * NUM_SUBTYPES_PER_TYPE;
     }
 }
